# Added ElementInfo caching for faster SvgElement Creation