--- a/Tests/Svg.UnitTests/PerformanceTest.cs
+++ b/Tests/Svg.UnitTests/PerformanceTest.cs
@@ -1,0 +1,45 @@
+﻿using System;
+using System.Collections.Generic;
+using System.IO;
+using System.Text;
+using System.Reflection;
+
+using NUnit.Framework;
+
+namespace Svg.UnitTests
+{
+    [TestFixture]
+    public class PerformanceTest 
+    {
+        public static string AssemblyDirectory
+        {
+            get
+            {
+                string codeBase = typeof(PerformanceTest).Assembly.CodeBase;
+                UriBuilder uri = new UriBuilder(codeBase);
+                string path = Uri.UnescapeDataString(uri.Path);
+                return Path.GetDirectoryName(path);
+            }
+        }
+
+        /// <summary>
+        /// Load All W3CSvg Unit tests
+        /// On My Machine this Unit Test went from 22 seconds down to 10 seconds. I set the Timeout
+        /// to 25 seconds to catch Performance regressions
+        /// </summary>
+        [Test]
+        [Timeout(25000)]
+        public void LoadAllW3CSvg()
+        {
+            var svgPath = Path.Combine(AssemblyDirectory, "..", "..", "..", "..", "W3CTestSuite", "svg");
+            var files = Directory.GetFiles(svgPath, "*.svg");
+            foreach (var file in files)
+            {
+                for (int i = 0; i < 10; i++)
+                {
+                    SvgDocument.Open<SvgDocument>(file);
+                }
+            }
+        }
+    }
+}