This causes the Performance Test to finish in 10 seconds instead of 22 seconds without the optimizations.

#### Reference Issue
https://github.com/vvvv/SVG/issues/767

#### What does this implement/fix? Explain your changes.
I profiled a Mobile App that uses this svg library and the profiler showed that a large part of the parsing of Svg is spent in 
creating the ElementInfo, I cache it statically, because it can't change during loading.

#### Any other comments?
Added an unit test for performance that times out after 25 seconds. Maybe the timeout can be adjusted downwards.
On my machine this Unit Test runs in 10 seconds.