@mrbean-bremen 2020-12-27
@H1Gdev, anyone else interested: this and the other performance PR seem to be ready to be merged, is there anything preventing the merge from your perspective?