@wieslawsoltes 2021-01-01
I am working on similar perf improvement in #772, Source Generators have advantage of being compile time operation, also remove usage of `Activator.CreateInstance` which should add even more performance.