@mrbean-bremen 2021-01-01
@wieslawsoltes: if you are adding to it, it would make sense to merge it (and the other performance PR) now, what do you think?