@wieslawsoltes 2021-01-01
> > I am working on similar perf improvement in #772
> 
> Would this replace this PR, or add to it?

I can add my changes after this PR