@wieslawsoltes 2021-01-01
> Ok, I merged the two performance PRs (and also a small fix I made).

Great, will update my PR